+++ src/test/java/com/readum/domain/user/userbook/service/UserBookCreateServiceTest.java
@@ -1,34 +1,25 @@
 package com.readum.domain.user.userbook.service;
 
 import com.readum.domain.book.dto.BookResult;
+import com.readum.domain.book.exception.BookErrorCode;
 import com.readum.domain.book.out.BookLookupClient;
-import com.readum.domain.exception.ConflictException;
+import com.readum.domain.exception.NotFoundException;
 import com.readum.domain.user.userbook.dto.UserBookCreateCommand;
 import com.readum.domain.user.userbook.dto.UserBookCreateResult;
-import com.readum.domain.user.userbook.exception.UserBookErrorCode;
-import com.readum.model.book.entity.Book;
-import com.readum.model.book.entity.BookFixture;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.entity.UserBookFixture;
-import com.readum.model.book.repository.BookRepository;
-import com.readum.model.user.repository.UserBookRepository;
-import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
+import org.mockito.InOrder;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.dao.DataIntegrityViolationException;
 
 import java.time.LocalDateTime;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 
 @ExtendWith(MockitoExtension.class)
 class UserBookCreateServiceTest {
@@ -37,19 +28,12 @@ class UserBookCreateServiceTest {
     private BookLookupClient bookLookupClient;
 
     @Mock
-    private BookRepository bookRepository;
-
-    @Mock
-    private UserBookRepository userBookRepository;
-
-    @Mock
-    private UserBookConflictReader userBookConflictReader;
+    private UserBookRegistrationWriter userBookRegistrationWriter;
 
     @InjectMocks
     private UserBookCreateService userBookCreateService;
 
     private static final Long USER_ID = 1L;
-    private static final Long BOOK_ID = 10L;
     private static final String EXTERNAL_ID = "9788965700807";
     private static final String TITLE = "테스트 책";
     private static final String AUTHORS = "테스트 저자";
@@ -65,62 +49,33 @@ class UserBookCreateServiceTest {
         return new BookResult(COVER_URL, TITLE, AUTHORS, PUBLISHER, PUBLISHED_YEAR, EXTERNAL_ID);
     }
 
-    private Book stubBook() {
-        return BookFixture.persistedBook(BOOK_ID, EXTERNAL_ID, TITLE, AUTHORS, PUBLISHER, PUBLISHED_YEAR, COVER_URL);
-    }
-
     @Test
-    void 신규_도서_등록_시_알라딘_조회_후_upsert_호출_후_UserBook이_저장되고_올바른_결과를_반환한다() {
-        Book book = stubBook();
-        LocalDateTime savedAt = LocalDateTime.of(2024, 6, 1, 12, 0);
-        UserBook savedUserBook = UserBookFixture.persistedUserBook(100L, USER_ID, BOOK_ID, savedAt);
+    void 도서_등록_시_알라딘_조회_후_그_결과로_DB_쓰기를_위임하고_쓰기_결과를_반환한다() {
+        BookResult bookInfo = stubBookResult();
+        UserBookCreateResult writerResult = new UserBookCreateResult(
+                100L, USER_ID, EXTERNAL_ID, TITLE, AUTHORS, PUBLISHER, PUBLISHED_YEAR, COVER_URL,
+                LocalDateTime.of(2024, 6, 1, 12, 0));
 
-        given(bookLookupClient.execute(EXTERNAL_ID)).willReturn(stubBookResult());
-        given(bookRepository.findByExternalId(EXTERNAL_ID)).willReturn(Optional.of(book));
-        given(userBookRepository.save(any(UserBook.class))).willReturn(savedUserBook);
+        given(bookLookupClient.execute(EXTERNAL_ID)).willReturn(bookInfo);
+        given(userBookRegistrationWriter.register(command(), bookInfo)).willReturn(writerResult);
 
         UserBookCreateResult result = userBookCreateService.execute(command());
 
-        verify(bookLookupClient).execute(EXTERNAL_ID);
-        verify(bookRepository).upsert(EXTERNAL_ID, TITLE, AUTHORS, PUBLISHER, PUBLISHED_YEAR, COVER_URL);
+        assertThat(result).isEqualTo(writerResult);
 
-        ArgumentCaptor<UserBook> userBookCaptor = ArgumentCaptor.forClass(UserBook.class);
-        verify(userBookRepository).save(userBookCaptor.capture());
-        assertThat(userBookCaptor.getValue().getUserId()).isEqualTo(USER_ID);
-        assertThat(userBookCaptor.getValue().getBookId()).isEqualTo(BOOK_ID);
-
-        assertThat(result.id()).isEqualTo(100L);
-        assertThat(result.userId()).isEqualTo(USER_ID);
-        assertThat(result.bookExternalId()).isEqualTo(EXTERNAL_ID);
-        assertThat(result.title()).isEqualTo(TITLE);
-        assertThat(result.authors()).isEqualTo(AUTHORS);
-        assertThat(result.createdAt()).isEqualTo(savedAt);
+        InOrder inOrder = Mockito.inOrder(bookLookupClient, userBookRegistrationWriter);
+        inOrder.verify(bookLookupClient).execute(EXTERNAL_ID);
+        inOrder.verify(userBookRegistrationWriter).register(command(), bookInfo);
     }
 
     @Test
-    void 이미_등록된_도서_재등록_시_ALREADY_EXISTS_ErrorCode와_기존_UserBook_데이터가_payload에_담긴_ConflictException이_발생한다() {
-        Book book = stubBook();
-        LocalDateTime existingCreatedAt = LocalDateTime.of(2024, 1, 1, 0, 0);
-        UserBook existingUserBook = UserBookFixture.persistedUserBook(99L, USER_ID, BOOK_ID, existingCreatedAt);
-
-        given(bookLookupClient.execute(EXTERNAL_ID)).willReturn(stubBookResult());
-        given(bookRepository.findByExternalId(EXTERNAL_ID)).willReturn(Optional.of(book));
-        given(userBookRepository.save(any(UserBook.class))).willThrow(new DataIntegrityViolationException("duplicate"));
-        given(userBookConflictReader.find(USER_ID, BOOK_ID)).willReturn(Optional.of(existingUserBook));
+    void 알라딘_조회가_실패하면_DB_쓰기를_시도하지_않고_예외가_그대로_전파된다() {
+        NotFoundException lookupFailure = new NotFoundException(BookErrorCode.LOOKUP_NOT_FOUND);
+        given(bookLookupClient.execute(EXTERNAL_ID)).willThrow(lookupFailure);
 
         assertThatThrownBy(() -> userBookCreateService.execute(command()))
-                .asInstanceOf(InstanceOfAssertFactories.type(ConflictException.class))
-                .satisfies(ex -> {
-                    assertThat(ex.getErrorCode()).isEqualTo(UserBookErrorCode.ALREADY_EXISTS);
-                    assertThat(ex.getPayload())
-                            .asInstanceOf(InstanceOfAssertFactories.type(UserBookCreateResult.class))
-                            .satisfies(payload -> {
-                                assertThat(payload.id()).isEqualTo(99L);
-                                assertThat(payload.userId()).isEqualTo(USER_ID);
-                                assertThat(payload.bookExternalId()).isEqualTo(EXTERNAL_ID);
-                                assertThat(payload.title()).isEqualTo(TITLE);
-                                assertThat(payload.createdAt()).isEqualTo(existingCreatedAt);
-                            });
-                });
+                .isSameAs(lookupFailure);
+
+        verifyNoInteractions(userBookRegistrationWriter);
     }
 }
